refactor(channels): extract helpers to satisfy complexity ratchet
- Split whatsapp-channel.ts into pipeline.ts (webhook ingress + outbound
  send helpers) plus a thin factory (525 -> 167 lines).
- Extract handler-worker loop body into processClaim/handleBeginFailure/
  handleTimeout/handleHandlerError (170-line fn -> sub-50-line helpers).
- Extract webhook-handler tryBegin handling into beginOrResolve.
- Trim oversized comments in email/teams channel factories.

No behavior change; types and call sites preserved.
