+++ packages/lib/channel-teams/src/teams-channel.ts
@@ -60,12 +60,8 @@ export type TeamsChannelAdapter = ChannelAdapter & {
 
 const TEAMS_CAPABILITIES: ChannelCapabilities = {
   text: true,
-  // formatOutbound() currently serializes only text blocks. Advertising
-  // images/files/buttons here without an end-to-end serializer would
-  // silently drop those blocks on the wire — a contract break for
-  // capability-driven callers. Flip these to true only when format.ts
-  // emits the corresponding Bot Framework attachment / suggested-action
-  // payloads.
+  // formatOutbound() serializes only text in v1; flip these to true only
+  // when format.ts emits real attachment / suggested-action payloads.
   images: false,
   files: false,
   buttons: false,
